refactor(studio): split service payload helpers
