+++ scripts/qa/check_file_sizes.py
@@ -54,7 +54,6 @@ LEGACY_LIMITS: Final = {
     "src/contexts/studio/application/ports/studio_repository.py": 543,
     "src/shared/infrastructure/config/settings.py": 523,
     "src/contexts/studio/application/services/ai_service.py": 383,
-    "src/contexts/studio/application/service_common.py": 354,
 }
